[refactor] tpl파일들 함수 이름과 파일이름 통일
