+++ FE/src/js/mock.js
@@ -20,6 +20,14 @@ export const mock = {
           createTime: "2020-04-07 14:24:23",
           modifiedTime: "2020-04-07 15:24:23",
         },
+        {
+          id: 3,
+          title: null,
+          content: "github 공부하기2",
+          author: "web2",
+          createTime: "2020-04-07 14:24:23",
+          modifiedTime: "2020-04-07 15:24:23",
+        },
       ],
       cardsCount: 2,
       valid: true,
